Changed enums and Time properties to be byte, instead of int. This decreases the size of the Schedule struct to be within the 16 byte guidline.

+++ Assignment6/src/UniversityCourseWork/Time.cs
@@ -2,8 +2,8 @@
 {
     public struct Time
     {
-        public int hour { get; }
-        public int minute { get; }
-        public int second { get; }
+        public byte hour { get; }
+        public byte minute { get; }
+        public byte second { get; }
     }
 }
